Rewrite LGraphAdapters to match Java and complete remaining AGENTS tasks
LGraphAdapters now supports the full 4-parameter adapt() matching Java:
transparent N/S edges, transparent comment nodes, and node filter.
Key fixes: LNodeAdapter returns empty edges (not port edges),
is_compound_node checks only COMPOUND_NODE (removed inside_self_loops),
LPortAdapter handles transparent N/S port dummy edges and self-loop
holder edges, LGraphAdapter delegates all properties via cloned
MapPropertyHolder. LabelAndNodeSizeProcessor updated to call
adapt(graph, true, true, Normal filter).

Also includes prior uncommitted work: cycle breaker fixes,
p2layers/p3order improvements, layout transferrer enhancements,
and NDC node dimension calculation updates.

Parity: 88/100 ok, 19 match, 69 drift, 1307 diffs, 12 timeout
(unchanged from baseline — drift is from p1/p2/p3 processors).

+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/mod.rs
@@ -95,8 +95,10 @@ impl Random {
     }
 
     pub fn next_float(&mut self) -> f64 {
-        let value = self.next(24) as u64;
-        (value as f64) / ((1u64 << 24) as f64)
+        let value = self.next(24);
+        // Match Java's nextFloat() precision: return 32-bit float converted to f64
+        // Java: return next(24) / ((float)(1 << 24));
+        ((value as f32) / ((1u32 << 24) as f32)) as f64
     }
 
     pub fn next_int(&mut self, bound: i32) -> i32 {
